Better structure in the `verifier` module
